update temperature to long double

+++ optlib/sched.cpp
@@ -2,14 +2,16 @@
 
 #include <cmath>
 
-ExpDecayScheduler::ExpDecayScheduler(double start, double step, double min)
+ExpDecayScheduler::ExpDecayScheduler(long double start, long double step,
+                                     long double min)
     : temp_(start), step_(step), min_(min) {}
 
 bool ExpDecayScheduler::ShouldShift(double from_score, double to_score) {
     if (to_score < from_score) {
         return true;
     }
-    double shift_probability = exp(-(from_score/to_score)/temp_);
+    long double shift_probability =
+        std::exp(-(static_cast<long double>(from_score) / to_score) / temp_);
     return distr_(rnd_) < shift_probability;
 }
